[APPC-2249] - Moved more preferences to new impressions repo.
