+++ app/src/test/java/com/asfoundation/wallet/poa/ProofOfAttentionServiceTest.java
@@ -8,10 +8,9 @@ import com.asfoundation.wallet.advertise.CampaignInteract;
 import com.asfoundation.wallet.billing.partners.AddressService;
 import com.asfoundation.wallet.entity.Wallet;
 import com.asfoundation.wallet.interact.AutoUpdateInteract;
-import com.asfoundation.wallet.interact.WalletCreatorInteract;
 import com.asfoundation.wallet.interact.FindDefaultWalletInteract;
 import com.asfoundation.wallet.repository.BdsBackEndWriter;
-import com.asfoundation.wallet.repository.PreferencesRepositoryType;
+import com.asfoundation.wallet.repository.ImpressionPreferencesRepositoryType;
 import com.asfoundation.wallet.repository.WalletNotFoundException;
 import com.asfoundation.wallet.service.Campaign;
 import com.asfoundation.wallet.service.CampaignService;
@@ -51,19 +50,17 @@ public class ProofOfAttentionServiceTest {
   @Mock CampaignService campaignService;
   @Mock HashCalculator hashCalculator;
   @Mock AddressService addressService;
-  @Mock WalletCreatorInteract walletInteract;
-  @Mock FindDefaultWalletInteract findDefaultWalletInteract;
   @Mock AdvertisingThrowableCodeMapper mapper;
   @Mock WalletService walletService;
-  @Mock PreferencesRepositoryType preferences;
+  @Mock ImpressionPreferencesRepositoryType preferences;
   @Mock AutoUpdateInteract autoUpdateInteract;
   private int chainId;
   private ProofOfAttentionService proofOfAttentionService;
   private MemoryCache<String, Proof> cache;
-  private int maxNumberProofComponents = 3;
+  private final int maxNumberProofComponents = 3;
   private long nonce;
-  private String packageName = "package";
-  private int versionCode = 0;
+  private final String packageName = "package";
+  private final int versionCode = 0;
   private TestScheduler testScheduler;
   private BehaviorSubject<Wallet> hasWallet;
   private String wallet;
